requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
fonttools==4.53.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.9.0
msgpack==1.0.8
numpy==1.26.4
packaging==24.0
pillow==10.3.0
prettytable==3.10.0
psycopg2-binary==2.9.10
pynvim==0.5.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
requests==2.32.3
scipy==1.13.0
six==1.16.0
tqdm==4.66.2
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
